requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.9.0
    # via openai
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
openai==1.74.0
    # via -r requirements.in
pathspec==0.12.1
    # via -r requirements.in
pillow==11.3.0
    # via -r requirements.in
prompt-toolkit==3.0.51
    # via -r requirements.in
pydantic==2.11.3
    # via openai
pydantic-core==2.33.1
    # via pydantic
pygments==2.19.1
    # via
    #   -r requirements.in
    #   rich
pypdf2==3.0.1
    # via -r requirements.in
pyperclip==1.9.0
    # via -r requirements.in
python-dotenv==1.1.0
    # via -r requirements.in
regex==2025.7.34
    # via tiktoken
requests==2.32.4
    # via
    #   -r requirements.in
    #   tiktoken
rich==14.0.0
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tiktoken==0.10.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.13.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.5.0
    # via requests
urwid==3.0.2
    # via -r requirements.in
wcwidth==0.2.13
    # via
    #   prompt-toolkit
    #   urwid
